requirements.txt
asgiref==3.11.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
crispy-bootstrap5==0.7
cryptography==46.0.5
Django==4.2.7
django-cors-headers==4.3.1
django-crispy-forms==2.1
djangorestframework==3.14.0
google-ai-generativelanguage==0.4.0
google-api-core==2.29.0
google-auth==2.48.0
google-generativeai==0.3.0
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.62.3
gunicorn==21.2.0
idna==3.11
Markdown==3.5.1
packaging==26.0
Pillow==10.1.0
proto-plus==1.27.1
protobuf==4.25.8
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
python-dotenv==1.0.0
pytz==2025.2
requests==2.31.0
rsa==4.9.1
sqlparse==0.5.5
stripe==7.8.0
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.6.0
